CP2L1-**Level One Complete**
correctly checks in multiple patients and adds them to front of patientList

+++ CProject2/health.h
@@ -51,10 +51,10 @@ extern void addHealthType( int patientID, int newType );
 extern Chartptr getChart( int patientID );
 extern CBuffptr getHealthType( int patientID, int healthType ); 
 extern void removePatient( int patientID );
-extern void addHealthReading( CBuffptr buffer, char* timestamp, float value );
+extern void addHealthReading( CBuffptr buffer, char* timestamp, int value );
 
 //own function declarations from health_util
-extern int read_line( int *id, char *time[], int *type, float *val);
+extern int read_line( int *id, char *time[], int *type, int *val);
 extern void printPatient(int id);
 extern void resetPatient(int id);
